agent_squad_sdlc/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-squad-sdlc"
version = "0.1.0"
description = "Multi-agent SDLC automation using AWS Agent Squad with Claude"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "SDLC Automation Team" }
]
keywords = ["agent", "sdlc", "automation", "claude", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "agent-squad[anthropic]>=0.1.0",
    "PyGithub>=2.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0",
    "PyJWT>=2.8.0",
    "boto3>=1.34.0",
    "click>=8.1.0",
    "uvicorn>=0.24.0",
    "fastapi>=0.104.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "moto>=4.2.0",
    "responses>=0.24.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
sdlc-agent = "agent_squad_sdlc.handlers.cli_handler:main"

[tool.hatch.build.targets.wheel]
packages = ["agent_squad_sdlc"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
